int -> size_t for num_threads

+++ include/RAJA/policy/cuda/kernel/CudaKernel.hpp
@@ -58,7 +58,7 @@ namespace RAJA
  * runtime.
  * Num_threads is 1024, which may not be appropriate for all kernels.
  */
-template <bool async0, size_t num_blocks, int num_threads>
+template <bool async0, size_t num_blocks, size_t num_threads>
 struct cuda_launch {};
 
 /*!
@@ -67,7 +67,7 @@ struct cuda_launch {};
  * If num_blocks is 0 then num_blocks is chosen at runtime.
  * Num_blocks is chosen to maximize the number of blocks running concurrently.
  */
-template <bool async0, size_t num_blocks, int num_threads>
+template <bool async0, size_t num_blocks, size_t num_threads>
 using cuda_explicit_launch = cuda_launch<async0, num_blocks, num_threads>;
 
 
@@ -76,7 +76,7 @@ using cuda_explicit_launch = cuda_launch<async0, num_blocks, num_threads>;
  * are determined by the CUDA occupancy calculator.
  * If num_threads is 0 then num_threads is chosen at runtime.
  */
-template <int num_threads0, bool async0>
+template <size_t num_threads0, bool async0>
 using cuda_occ_calc_launch = cuda_launch<async0, 0, num_threads0>;
 
 namespace statement
@@ -99,7 +99,7 @@ struct CudaKernelExt
  * calculator determine the unspecified values.
  * The kernel launch is synchronous.
  */
-template <size_t num_blocks, int num_threads, typename... EnclosedStmts>
+template <size_t num_blocks, size_t num_threads, typename... EnclosedStmts>
 using CudaKernelExp =
     CudaKernelExt<cuda_launch<false, num_blocks, num_threads>, EnclosedStmts...>;
 
@@ -109,7 +109,7 @@ using CudaKernelExp =
  * calculator determine the unspecified values.
  * The kernel launch is asynchronous.
  */
-template <size_t num_blocks, int num_threads, typename... EnclosedStmts>
+template <size_t num_blocks, size_t num_threads, typename... EnclosedStmts>
 using CudaKernelExpAsync =
     CudaKernelExt<cuda_launch<true, num_blocks, num_threads>, EnclosedStmts...>;
 
@@ -136,7 +136,7 @@ using CudaKernelOccAsync =
  * number of threads (specified by num_threads)
  * The kernel launch is synchronous.
  */
-template <int num_threads, typename... EnclosedStmts>
+template <size_t num_threads, typename... EnclosedStmts>
 using CudaKernelFixed =
     CudaKernelExt<cuda_explicit_launch<false, operators::limits<size_t>::max(), num_threads>,
                   EnclosedStmts...>;
@@ -146,7 +146,7 @@ using CudaKernelFixed =
  * number of threads (specified by num_threads)
  * The kernel launch is asynchronous.
  */
-template <int num_threads, typename... EnclosedStmts>
+template <size_t num_threads, typename... EnclosedStmts>
 using CudaKernelFixedAsync =
     CudaKernelExt<cuda_explicit_launch<true, operators::limits<size_t>::max(), num_threads>,
                   EnclosedStmts...>;
@@ -252,7 +252,7 @@ struct CudaLaunchHelper;
  * The user may specify the number of threads and blocks or let one or both be
  * determined at runtime using the CUDA occupancy calculator.
  */
-template<bool async0, size_t num_blocks, int num_threads, typename StmtList, typename Data>
+template<bool async0, size_t num_blocks, size_t num_threads, typename StmtList, typename Data>
 struct CudaLaunchHelper<cuda_launch<async0, num_blocks, num_threads>,StmtList,Data>
 {
   using Self = CudaLaunchHelper;
@@ -264,7 +264,7 @@ struct CudaLaunchHelper<cuda_launch<async0, num_blocks, num_threads>,StmtList,Da
   using kernelGetter_t = CudaKernelLauncherGetter<(num_threads <= 0) ? 0 : num_threads, Data, executor_t>;
 
   inline static void recommended_blocks_threads(int shmem_size,
-      size_t &recommended_blocks, int &recommended_threads)
+      size_t &recommended_blocks, size_t &recommended_threads)
   {
     auto func = kernelGetter_t::get();
 
@@ -319,7 +319,7 @@ struct CudaLaunchHelper<cuda_launch<async0, num_blocks, num_threads>,StmtList,Da
     }
   }
 
-  inline static void max_threads(int RAJA_UNUSED_ARG(shmem_size), int &max_threads)
+  inline static void max_threads(int RAJA_UNUSED_ARG(shmem_size), size_t &max_threads)
   {
     if (num_threads <= 0) {
 
@@ -460,7 +460,7 @@ struct StatementExecutor<
 
     // Only launch kernel if we have something to iterate over
     size_t num_blocks = launch_dims.num_blocks();
-    int num_threads = launch_dims.num_threads();
+    size_t num_threads = launch_dims.num_threads();
     if (num_blocks > 0 || num_threads > 0) {
 
       //
@@ -474,7 +474,7 @@ struct StatementExecutor<
       // Compute the recommended physical kernel blocks and threads
       //
       size_t recommended_blocks;
-      int recommended_threads;
+      size_t recommended_threads;
       launch_t::recommended_blocks_threads(
           shmem, recommended_blocks, recommended_threads);
 
@@ -482,7 +482,7 @@ struct StatementExecutor<
       //
       // Compute the MAX physical kernel threads
       //
-      int max_threads;
+      size_t max_threads;
       launch_t::max_threads(shmem, max_threads);
